refactor : 유저, API 기능 따로 패키지에 정리
