+++ src/main/java/com/zerobase/whataday/schedule/domain/Schedule.java
@@ -1,7 +1,7 @@
 package com.zerobase.whataday.schedule.domain;
 
 import com.zerobase.whataday.domain.BaseEntity;
-import com.zerobase.whataday.domain.User;
+import com.zerobase.whataday.user.domain.User;
 import java.time.LocalDateTime;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
