--- src/main/java/com/zerobase/whataday/service/UserService.java
@@ -1,10 +1,0 @@
-package com.zerobase.whataday.service;
-
-import com.zerobase.whataday.domain.UserInput;
-import org.springframework.stereotype.Service;
-
-@Service
-public interface UserService {
-  String signUp(UserInput form);
-
-}
