+++ src/main/java/com/zerobase/whataday/user/controller/UserController.java
@@ -1,7 +1,7 @@
 package com.zerobase.whataday.user.controller;
 
-import com.zerobase.whataday.domain.User;
-import com.zerobase.whataday.domain.UserInput;
+import com.zerobase.whataday.user.domain.User;
+import com.zerobase.whataday.user.domain.UserRequest;
 import javax.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -12,15 +12,15 @@ import org.springframework.web.bind.annotation.RestController;
 public class UserController {
 
   @PostMapping("/user/submit")
-  public ResponseEntity<?> addUser(@RequestBody @Valid UserInput userInput) {
+  public ResponseEntity<?> addUser(@RequestBody @Valid UserRequest userRequest) {
 
 
 
 
     User user = User.builder()
-        .name(userInput.getName())
-        .password(userInput.getPassword())
-        .email(userInput.getEmail())
+        .name(userRequest.getName())
+        .password(userRequest.getPassword())
+        .email(userRequest.getEmail())
         .build();
 
     return ResponseEntity.ok().build();
