+++ src/main/java/com/zerobase/whataday/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.zerobase.whataday.repository;
+package com.zerobase.whataday.user.repository;
 
-import com.zerobase.whataday.domain.User;
+import com.zerobase.whataday.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
